+++ CHANGELOG.md
@@ -6,16 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).  
 
-## [0.1.1] - 2024-04-16  
+## Unreleased (To be deployed)  
 
 ### Added  
 - Roxygen comments to R util files.  
 - `utils_dt_display.R` to handle post-pooltestr processing.  
 - Prevalence division option (default 1/2000).  
+- CITATION file.  
 
 ### Changed  
 - Refactored pooltestr-related code.  
-- Analyse prevalence rounding now part of `utils_dt_display.R`  
+- Analyse prevalence rounding now part of `utils_dt_display.R`.  
+- About page loads from a .md.  
 
 ### Removed  
 - Redundant analyse datatable UI.